worker/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
dnspython==2.7.0
Flask==3.1.1
idna==3.10
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
netmiko==4.6.0
ntc_templates==8.0.0
paramiko==4.0.0
pika==1.3.2
pycparser==2.22
Pygments==2.19.2
pymongo==4.14.0
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scp==0.15.0
textfsm==2.1.0
urllib3==2.5.0
Werkzeug==3.1.3
